requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
authlib==1.4.0
httpx==0.28.1
itsdangerous==2.2.0
python-dotenv==1.0.1
pydantic==2.10.3
jinja2==3.1.4
python-multipart==0.0.20
Pillow>=10.0.0
pytest>=8.0.0
